smite-scenarios: split mocks and test helpers into submodules
Program-building helpers are moved to tests/programs.rs.

Mocks and other helper functions are moved to tests/harness.rs.

The reason for this separation will become clearer in subsequent PRs
where we will significantly refactor and extend the existing
program-builders and helpers.

Two decoder helpers and recv_channel_ready_executor are left in tests.rs
and will be deleted entirely in subsequent PRs.
