+++ smite-scenarios/src/executor/tests.rs
@@ -1,367 +1,16 @@
-use std::collections::VecDeque;
+mod harness;
+mod programs;
+
 use std::str::FromStr;
 
 use super::*;
+use bitcoin::Amount;
 use bitcoin::secp256k1::{Secp256k1, SecretKey};
-use bitcoin::{Amount, Transaction};
+use harness::*;
+use programs::*;
 use smite::bolt::{AcceptChannelTlvs, GossipTimestampFilter, Init, Ping};
 use smite_ir::Instruction;
 use smite_ir::operation::ShutdownScriptVariant;
-
-// -- MockConnection --
-
-struct MockConnection {
-    recv_queue: VecDeque<Vec<u8>>,
-    sent: Vec<Vec<u8>>,
-}
-
-impl MockConnection {
-    fn new() -> Self {
-        Self {
-            recv_queue: VecDeque::new(),
-            sent: Vec::new(),
-        }
-    }
-
-    fn queue_recv(&mut self, msg_bytes: Vec<u8>) {
-        self.recv_queue.push_back(msg_bytes);
-    }
-}
-
-impl Connection for MockConnection {
-    fn send_message(&mut self, msg: &[u8]) -> Result<(), ConnectionError> {
-        self.sent.push(msg.to_vec());
-        Ok(())
-    }
-
-    fn recv_message(&mut self) -> Result<Vec<u8>, ConnectionError> {
-        self.recv_queue
-            .pop_front()
-            .ok_or_else(|| ConnectionError::Io(std::io::ErrorKind::UnexpectedEof.into()))
-    }
-
-    fn set_read_timeout(&mut self, _timeout: Option<Duration>) -> Result<(), ConnectionError> {
-        Ok(())
-    }
-
-    fn read_timeout(&self) -> Result<Option<Duration>, ConnectionError> {
-        Ok(None)
-    }
-}
-
-// Mocking BitcoinCli via MockBitcoinCli
-
-#[derive(Default)]
-struct MockBitcoinCli {
-    mine_blocks_calls: Vec<u8>,
-    mined_private_mempool: Vec<String>,
-    broadcast_calls: Vec<Transaction>,
-    block_position_lookups: Vec<Txid>,
-    utxos: Vec<Utxo>,
-    change_spk: ScriptBuf,
-    confirmations: u32,
-}
-
-impl BitcoinRpc for MockBitcoinCli {
-    fn mine_blocks(&mut self, num_blocks: u8, private_mempool: &[String]) {
-        self.mine_blocks_calls.push(num_blocks);
-        self.mined_private_mempool = private_mempool.to_vec();
-        self.confirmations += u32::from(num_blocks);
-    }
-
-    fn get_utxos(&mut self) -> Vec<Utxo> {
-        self.utxos.clone()
-    }
-
-    fn get_new_address_script_pubkey(&mut self) -> ScriptBuf {
-        self.change_spk.clone()
-    }
-
-    fn sign_and_broadcast_tx(&mut self, tx: &bitcoin::Transaction) -> Option<String> {
-        self.broadcast_calls.push(tx.clone());
-
-        // Simulate a mempool-policy rejection: if any output is below its
-        // dust threshold, return the tx hex so it gets queued in the
-        // private mempool. Otherwise the tx is accepted.
-        let has_dust = tx
-            .output
-            .iter()
-            .any(|o| o.value < o.script_pubkey.minimal_non_dust());
-        if has_dust {
-            Some(bitcoin::consensus::encode::serialize_hex(tx))
-        } else {
-            None
-        }
-    }
-
-    fn lock_utxos(&mut self, outpoints: &[OutPoint]) {
-        self.utxos.retain(|u| !outpoints.contains(&u.outpoint));
-    }
-
-    fn get_transaction_confirmations(&mut self, _txid: Txid) -> u32 {
-        self.confirmations
-    }
-
-    fn get_transaction_block_position(&mut self, txid: Txid) -> Option<TxBlockPosition> {
-        self.block_position_lookups.push(txid);
-        // Distinctive coordinates so tests can verify the executor
-        // combined them with the funding transaction's vout.
-        (self.confirmations > 0).then_some(TxBlockPosition {
-            block_height: 800_042,
-            tx_index: 7,
-        })
-    }
-}
-
-// -- Helpers --
-
-fn sample_pubkey(byte: u8) -> PublicKey {
-    let secp = Secp256k1::new();
-    let mut key_bytes = [0u8; 32];
-    key_bytes[31] = byte;
-    let sk = SecretKey::from_slice(&key_bytes).expect("valid secret key");
-    PublicKey::from_secret_key(&secp, &sk)
-}
-
-fn sample_context() -> ProgramContext {
-    ProgramContext {
-        target_pubkey: sample_pubkey(1),
-        chain_hash: [0xcc; 32],
-        block_height: 800_000,
-        target_features: vec![],
-    }
-}
-
-fn sample_utxo() -> Utxo {
-    Utxo {
-        amount: Amount::from_sat(10_008_942),
-        outpoint: OutPoint {
-            txid: "a1f7b953dc8c3db0222d931d3e2613f9971af75a09a005b31af057f8414cc5d7"
-                .parse()
-                .expect("valid txid"),
-            vout: 0,
-        },
-        script_pubkey: ScriptBuf::from(
-            hex::decode("0014a10d9257489e685dda030662390dc177852faf13")
-                .expect("valid P2WPKH scriptpubkey hex"),
-        ),
-    }
-}
-
-fn sample_change_spk() -> ScriptBuf {
-    ScriptBuf::from(
-        hex::decode("00142e532c12351a5c81e23c8a76d19345ca7b6de57a")
-            .expect("valid P2WPKH scriptpubkey hex"),
-    )
-}
-
-fn sample_accept_channel() -> AcceptChannel {
-    AcceptChannel {
-        temporary_channel_id: TemporaryChannelId::new([0xbb; 32]),
-        dust_limit_satoshis: 546,
-        max_htlc_value_in_flight_msat: 100_000_000,
-        channel_reserve_satoshis: 10_000,
-        htlc_minimum_msat: 1_000,
-        minimum_depth: 6,
-        to_self_delay: 144,
-        max_accepted_htlcs: 483,
-        funding_pubkey: sample_pubkey(1),
-        revocation_basepoint: sample_pubkey(2),
-        payment_basepoint: sample_pubkey(3),
-        delayed_payment_basepoint: sample_pubkey(4),
-        htlc_basepoint: sample_pubkey(5),
-        first_per_commitment_point: sample_pubkey(6),
-        tlvs: AcceptChannelTlvs {
-            upfront_shutdown_script: Some(vec![0xde, 0xad]),
-            channel_type: Some(vec![0x40, 0x10, 0x00]),
-        },
-    }
-}
-
-/// Builds the 20 `open_channel` input instructions in wire order.
-fn open_channel_instructions() -> Vec<Instruction> {
-    vec![
-        Instruction {
-            operation: Operation::LoadChainHashFromContext,
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadChannelId([0xbb; 32]),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadAmount(100_000),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadAmount(0),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadAmount(546),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadAmount(100_000_000),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadAmount(10_000),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadAmount(1_000),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadFeeratePerKw(253),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadU16(144),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadU16(483),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadTargetPubkeyFromContext,
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadTargetPubkeyFromContext,
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadTargetPubkeyFromContext,
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadTargetPubkeyFromContext,
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadTargetPubkeyFromContext,
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadTargetPubkeyFromContext,
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadU8(1),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadBytes(vec![]),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadFeatures(vec![0x40, 0x10, 0x00]),
-            inputs: vec![],
-        },
-    ]
-}
-
-fn create_and_broadcast_tx_instructions() -> Vec<Instruction> {
-    let opener_privkey =
-        SecretKey::from_str("30ff4956bbdd3222d44cc5e8a1261dab1e07957bdac5ae88fe3261ef321f3749")
-            .unwrap()
-            .secret_bytes();
-    let acceptor_privkey =
-        SecretKey::from_str("1552dfba4f6cf29a62a0af13c8d6981d36d0ef8d61ba10fb0fe90da7634d7e13")
-            .unwrap()
-            .secret_bytes();
-
-    vec![
-        Instruction {
-            operation: Operation::LoadPrivateKey(opener_privkey),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::DerivePoint,
-            inputs: vec![0],
-        },
-        Instruction {
-            operation: Operation::LoadPrivateKey(acceptor_privkey),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::DerivePoint,
-            inputs: vec![2],
-        },
-        Instruction {
-            operation: Operation::LoadAmount(10_000_000),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadFeeratePerKw(15_000),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::CreateFundingTransaction,
-            inputs: vec![1, 3, 4, 5],
-        },
-        Instruction {
-            operation: Operation::BroadcastTransaction,
-            inputs: vec![6],
-        },
-    ]
-}
-
-/// Builds instructions that construct and send a `channel_announcement`
-/// referencing the `ShortChannelId` produced at variable index `scid_var`.
-///
-/// `base` is the variable index the first appended instruction will occupy
-/// (i.e. the current program length), used to wire up the inputs to
-/// `BuildChannelAnnouncement`.
-fn channel_announcement_from_scid_instructions(base: usize, scid_var: usize) -> Vec<Instruction> {
-    vec![
-        Instruction {
-            operation: Operation::LoadFeatures(vec![0x01, 0x02]),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadChainHashFromContext,
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadPrivateKey([0x11; 32]),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadPrivateKey([0x22; 32]),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadPrivateKey([0x33; 32]),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::LoadPrivateKey([0x44; 32]),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::BuildChannelAnnouncement,
-            // features, chain_hash, short_channel_id, node_sk_1, node_sk_2,
-            // bitcoin_sk_1, bitcoin_sk_2.
-            inputs: vec![
-                base,
-                base + 1,
-                scid_var,
-                base + 2,
-                base + 3,
-                base + 4,
-                base + 5,
-            ],
-        },
-        Instruction {
-            operation: Operation::SendMessage,
-            inputs: vec![base + 6],
-        },
-    ]
-}
 
 /// Decodes a sent message expected to be a `channel_announcement`.
 fn decode_sent_channel_announcement(bytes: &[u8]) -> ChannelAnnouncement {
@@ -377,22 +26,6 @@ fn decode_open_channel(bytes: &[u8]) -> OpenChannel {
         other => panic!("expected open_channel(32), got {other}"),
     }
 }
-
-fn send_open_channel_instructions() -> Vec<Instruction> {
-    let mut instructions = open_channel_instructions();
-    instructions.extend([
-        Instruction {
-            operation: Operation::BuildOpenChannel,
-            inputs: (0..20).collect(),
-        },
-        Instruction {
-            operation: Operation::SendOpenChannel,
-            inputs: vec![20],
-        },
-    ]);
-    instructions
-}
-
 // -- execute() tests --
 
 #[test]
@@ -1839,95 +1472,6 @@ fn execute_create_funding_transaction_insufficient_funds() {
     };
     assert_eq!(funds_err.available, Amount::from_sat(1_000));
     assert_eq!(funds_err.required, Amount::from_sat(10_007_290));
-}
-
-#[allow(clippy::similar_names)]
-fn sample_funding_negotiation() -> PendingChannel {
-    let secp = Secp256k1::new();
-    let opener_sk =
-        SecretKey::from_str("30ff4956bbdd3222d44cc5e8a1261dab1e07957bdac5ae88fe3261ef321f3749")
-            .unwrap();
-    let acceptor_sk =
-        SecretKey::from_str("1552dfba4f6cf29a62a0af13c8d6981d36d0ef8d61ba10fb0fe90da7634d7e13")
-            .unwrap();
-    let opener_pk = PublicKey::from_secret_key(&secp, &opener_sk);
-    let acceptor_pk = PublicKey::from_secret_key(&secp, &acceptor_sk);
-
-    PendingChannel {
-        open_channel: OpenChannel {
-            chain_hash: [0xcc; 32],
-            temporary_channel_id: TemporaryChannelId::new([0xbb; 32]),
-            funding_satoshis: 10_000_000,
-            push_msat: 3_000_000_000,
-            dust_limit_satoshis: 546,
-            max_htlc_value_in_flight_msat: 100_000_000,
-            channel_reserve_satoshis: 10_000,
-            htlc_minimum_msat: 1_000,
-            feerate_per_kw: 15_000,
-            to_self_delay: 144,
-            max_accepted_htlcs: 483,
-            funding_pubkey: opener_pk,
-            revocation_basepoint: opener_pk,
-            payment_basepoint: opener_pk,
-            delayed_payment_basepoint: opener_pk,
-            htlc_basepoint: opener_pk,
-            first_per_commitment_point: opener_pk,
-            channel_flags: 1,
-            tlvs: OpenChannelTlvs::default(),
-        },
-        accept_channel: Some(AcceptChannel {
-            temporary_channel_id: TemporaryChannelId::new([0xbb; 32]),
-            dust_limit_satoshis: 546,
-            max_htlc_value_in_flight_msat: 100_000_000,
-            channel_reserve_satoshis: 10_000,
-            htlc_minimum_msat: 1_000,
-            minimum_depth: 6,
-            to_self_delay: 144,
-            max_accepted_htlcs: 483,
-            funding_pubkey: acceptor_pk,
-            revocation_basepoint: acceptor_pk,
-            payment_basepoint: acceptor_pk,
-            delayed_payment_basepoint: acceptor_pk,
-            htlc_basepoint: acceptor_pk,
-            first_per_commitment_point: acceptor_pk,
-            tlvs: AcceptChannelTlvs::default(),
-        }),
-        funding_built: false,
-    }
-}
-
-fn send_funding_created_and_recv_funding_signed_instructions() -> Vec<Instruction> {
-    let mut instrs = create_and_broadcast_tx_instructions();
-    instrs.extend(vec![
-        Instruction {
-            operation: Operation::LoadChannelId([0xbb; 32]),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::SendFundingCreated,
-            inputs: vec![6, 0, 8],
-        },
-        Instruction {
-            operation: Operation::RecvFundingSigned,
-            inputs: vec![9],
-        },
-    ]);
-    instrs
-}
-
-fn recv_channel_ready_instructions(confirmations: u8) -> Vec<Instruction> {
-    let mut instrs = send_funding_created_and_recv_funding_signed_instructions();
-    instrs.extend([
-        Instruction {
-            operation: Operation::MineBlocks(confirmations),
-            inputs: vec![],
-        },
-        Instruction {
-            operation: Operation::RecvChannelReady,
-            inputs: vec![],
-        },
-    ]);
-    instrs
 }
 
 #[test]
